+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.6.25
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.16
+  - @ballerine/workflow-core@0.6.25
+
 ## 0.6.24
 
 ### Patch Changes